Student-Engagement-main 3/requirements.txt
opencv-python==4.9.0.80
keras==2.11.0
numpy==1.21.6
pandas==1.3.5
cmake==3.28.3
dlib==19.24.2
tqdm==4.66.2
PyYAML==6.0.1
